# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for hotel actions",
        "title": "Hotel Booking API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/Create": {
            "post": {
                "description": "create room",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "room"
                ],
                "summary": "create room",
                "operationId": "createRoom",
                "parameters": [
                    {
                        "description": "new room data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Room"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreatingRoomResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON or validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict (room already exists)",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "Request entity too large",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "415": {
                        "description": "Unsupported media type",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/GetFilteredRooms": {
            "post": {
                "description": "get list of rooms with optional filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "room"
                ],
                "summary": "get filtered rooms",
                "operationId": "getFilteredRooms",
                "parameters": [
                    {
                        "description": "filter criteria",
                        "name": "input",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "list of rooms",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Room"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid JSON or validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/Patch": {
            "patch": {
                "description": "patch an existing room",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "room"
                ],
                "summary": "patch room",
                "operationId": "patchRoom",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "room id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "patch data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RoomPatch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "rooms updated",
                        "schema": {
                            "$ref": "#/definitions/dto.RoomPatchResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON or validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/RemoveRoom": {
            "delete": {
                "description": "remove an existing room by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "room"
                ],
                "summary": "remove room",
                "operationId": "removeRoom",
                "parameters": [
                    {
                        "description": "room id to remove",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RemoveRoomRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Removed Room id: {id}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON or validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.CreatingRoomResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "roomId": {
                    "type": "integer"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "dto.RemoveRoomRequest": {
            "type": "object",
            "properties": {
                "roomId": {
                    "type": "integer"
                }
            }
        },
        "dto.RoomPatch": {
            "type": "object",
            "properties": {
                "floor": {
                    "type": "integer",
                    "example": 2
                },
                "isOccupied": {
                    "type": "boolean",
                    "example": true
                },
                "needCleaning": {
                    "type": "boolean",
                    "example": true
                },
                "roomCount": {
                    "type": "integer",
                    "example": 2
                },
                "roomType": {
                    "type": "string",
                    "example": "luxury"
                },
                "sleepingPlaces": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "dto.RoomPatchResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
        "model.Room": {
            "type": "object",
            "properties": {
                "floor": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_occupied": {
                    "type": "boolean"
                },
                "need_cleaning": {
                    "type": "boolean"
                },
                "number": {
                    "type": "integer"
                },
                "room_count": {
                    "type": "integer"
                },
                "room_type": {
                    "type": "string"
                },
                "sleeping_places": {
                    "type": "integer"
                }
            }
        }
    }
}